requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.7.0
    # via -r requirements.in
aiohttp==3.8.3
    # via discord.py
aiosignal==1.3.1
    # via aiohttp
anyio==3.5.0
    # via
    #   httpcore
    #   starlette
arrow==1.1.1
    # via withings-api
asgiref==3.4.1
    # via uvicorn
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via aiohttp
azure-core==1.22.0
    # via
    #   azure-identity
    #   azure-storage-blob
azure-identity==1.7.1
    # via -r requirements.in
azure-storage-blob==12.9.0
    # via -r requirements.in
cachetools==5.3.0
    # via google-auth
certifi==2021.5.30
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.4
    # via
    #   aiohttp
    #   httpx
    #   requests
click==8.0.1
    # via uvicorn
colorhash==1.0.4
    # via -r requirements.in
cryptography==3.4.8
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
    #   python-jose
discord.py==2.1.0
    # via -r requirements.in
dropbox==11.36.0
    # via -r requirements.in
ecdsa==0.17.0
    # via python-jose
fastapi==0.89.1
    # via
    #   -r requirements.in
    #   fastapi-jwt-auth
fastapi-jwt-auth==0.5.0
    # via -r requirements.in
filelock==3.9.0
    # via -r requirements.in
fitbit==0.3.1
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
geographiclib==1.52
    # via geopy
geopy==2.2.0
    # via -r requirements.in
google-api-core==2.11.0
    # via google-api-python-client
google-api-python-client==2.77.0
    # via -r requirements.in
google-auth==2.16.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==1.0.0
    # via -r requirements.in
googleapis-common-protos==1.58.0
    # via google-api-core
googlemaps==4.6.0
    # via -r requirements.in
gpxpy==1.5.0
    # via -r requirements.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.14.7
    # via httpx
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.22.0
    # via -r requirements.in
idna==3.2
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
isodate==0.6.1
    # via msrest
msal==1.16.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==0.3.1
    # via azure-identity
msrest==0.6.21
    # via azure-storage-blob
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.24.1
    # via
    #   pandas
    #   pyarrow
oauthlib==3.1.1
    # via requests-oauthlib
openrouteservice==2.3.3
    # via -r requirements.in
pandas==1.5.3
    # via -r requirements.in
pendulum==2.1.2
    # via -r requirements.in
pillow==9.0.1
    # via
    #   -r requirements.in
    #   staticmap
ply==3.11
    # via stone
polar-accesslink==0.0.5
    # via -r requirements.in
polyline==2.0.0
    # via -r requirements.in
portalocker==2.3.2
    # via msal-extensions
protobuf==4.21.12
    # via
    #   google-api-core
    #   googleapis-common-protos
pyarrow==10.0.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via
    #   -r requirements.in
    #   fastapi
    #   withings-api
pyjwt[crypto]==1.7.1
    # via
    #   fastapi-jwt-auth
    #   msal
pyparsing==3.0.9
    # via httplib2
python-dateutil==2.8.2
    # via
    #   arrow
    #   fitbit
    #   pandas
    #   pendulum
python-dotenv==0.19.0
    # via -r requirements.in
python-jose[cryptography]==3.3.0
    # via -r requirements.in
pytz==2022.7.1
    # via pandas
pytzdata==2020.1
    # via pendulum
requests==2.26.0
    # via
    #   azure-core
    #   dropbox
    #   google-api-core
    #   googlemaps
    #   msal
    #   msrest
    #   openrouteservice
    #   polar-accesslink
    #   requests-oauth
    #   requests-oauthlib
    #   staticmap
requests-oauth==0.4.1
    # via withings-api
requests-oauthlib==1.3.1
    # via
    #   fitbit
    #   google-auth-oauthlib
    #   msrest
    #   withings-api
rfc3986[idna2008]==1.5.0
    # via httpx
rsa==4.7.2
    # via
    #   google-auth
    #   python-jose
schedule==1.1.0
    # via -r requirements.in
six==1.16.0
    # via
    #   azure-core
    #   azure-identity
    #   dropbox
    #   ecdsa
    #   google-auth
    #   google-auth-httplib2
    #   isodate
    #   python-dateutil
    #   stone
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.22.0
    # via fastapi
staticmap==0.5.5
    # via -r requirements.in
stone==3.3.1
    # via dropbox
typing-extensions==4.4.0
    # via
    #   pydantic
    #   starlette
    #   withings-api
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.6
    # via requests
uvicorn==0.14.0
    # via -r requirements.in
withings-api==2.4.0
    # via -r requirements.in
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
